requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.1
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
pymemcache==3.0.1
python-dateutil==2.8.1
pytz==2019.3
rake-nltk==1.0.4
six==1.14.0
slackclient==2.5.0
Werkzeug==1.0.0
yarl==1.4.2
